feat: add translate quotes and requirements.txt
