+++ citacoes_diarias.json
@@ -1,0 +1,10 @@
+{
+    "data": "2025-01-03",
+    "citacoes": {
+        "Life": {
+            "texto": "I like the dreams of the future better than the history of the past.",
+            "autor": "Thomas Jefferson",
+            "genero": "Life"
+        }
+    }
+}
