+++ favoritos_citacoes.json
@@ -1,0 +1,10 @@
+{
+    "favoritos": [
+        {
+            "texto": "Getting lost along your path is a part of finding the path you are meant to be on.",
+            "autor": "Robin Sharma",
+            "genero": "Life",
+            "timestamp": "2025-01-03 14:38:32"
+        }
+    ]
+}
