+++ historico_citacoes.json
@@ -1,0 +1,130 @@
+{
+    "historico": [
+        {
+            "texto": "If you don't know where you're going, you will probably end up somewhere else.",
+            "autor": "Laurence J. Peter",
+            "genero": "Life",
+            "timestamp": "2025-01-03 16:31:44"
+        },
+        {
+            "texto": "Monsters are real, and ghosts are real too. They live inside us, and sometimes, they win.",
+            "autor": "Stephen King",
+            "genero": "Life",
+            "timestamp": "2025-01-03 16:30:28"
+        },
+        {
+            "texto": "When we think about our strengths, we are strong. When we think about our weaknesses, we are weak.",
+            "autor": "Peter A. Cohen",
+            "genero": "Life",
+            "timestamp": "2025-01-03 16:09:48"
+        },
+        {
+            "texto": "It's not how much we give but how much love we put into giving.",
+            "autor": "Mother Teresa",
+            "genero": "Life",
+            "timestamp": "2025-01-03 16:04:04"
+        },
+        {
+            "texto": "Life is simply what our feelings do to us.",
+            "autor": "Honore de Balzac",
+            "genero": "Life",
+            "timestamp": "2025-01-03 16:03:16"
+        },
+        {
+            "texto": "Life is growth. If we stop growing, technically and spiritually, we are as good as dead.",
+            "autor": "Morihei Ueshiba",
+            "genero": "Life",
+            "timestamp": "2025-01-03 15:51:35"
+        },
+        {
+            "texto": "If you don't read the newspaper, you're uninformed. If you read the newspaper, you're misinformed.",
+            "autor": "Mark Twain",
+            "genero": "Life",
+            "timestamp": "2025-01-03 15:39:50"
+        },
+        {
+            "texto": "There is no limitations except those we create for ourselves.",
+            "autor": "Lolly Daskal",
+            "genero": "Life",
+            "timestamp": "2025-01-03 15:39:20"
+        },
+        {
+            "texto": "When uncomfortable, my instinct is not to avoid the discomfort but to become at peace with it.",
+            "autor": "Josh Waitzkin",
+            "genero": "Life",
+            "timestamp": "2025-01-03 15:39:12"
+        },
+        {
+            "texto": "The more you are grateful for what you have the more you will have to be grateful for.  ",
+            "autor": "Zig Ziglar",
+            "genero": "Life",
+            "timestamp": "2025-01-03 15:39:06"
+        },
+        {
+            "texto": "The spirit is beyond destruction. No one can bring an end to spirit which is everlasting.",
+            "autor": "Bhagavad Gita",
+            "genero": "Life",
+            "timestamp": "2025-01-03 15:22:11"
+        },
+        {
+            "texto": "Don't let what you cannot do interfere with what you can do.",
+            "autor": "John Wooden",
+            "genero": "Life",
+            "timestamp": "2025-01-03 15:19:56"
+        },
+        {
+            "texto": "We would accomplish many more things if we did not think of them as impossible.",
+            "autor": "Vince Lombardi",
+            "genero": "Life",
+            "timestamp": "2025-01-03 15:19:49"
+        },
+        {
+            "texto": "Men go to far greater lengths to avoid what they fear than to obtain what they desire.",
+            "autor": "Dan Brown",
+            "genero": "Life",
+            "timestamp": "2025-01-03 14:38:39"
+        },
+        {
+            "texto": "Getting lost along your path is a part of finding the path you are meant to be on.",
+            "autor": "Robin Sharma",
+            "genero": "Life",
+            "timestamp": "2025-01-03 14:38:31"
+        },
+        {
+            "texto": "Happiness is a gift and the trick is not to expect it, but to delight in it when it comes.",
+            "autor": "Charles Dickens",
+            "genero": "Life",
+            "timestamp": "2025-01-03 14:35:29"
+        },
+        {
+            "texto": "Success is liking yourself, liking what you do, and liking how you do it.",
+            "autor": "Maya Angelou",
+            "genero": "Life",
+            "timestamp": "2025-01-03 14:35:23"
+        },
+        {
+            "texto": "A gentleman is one who puts more into the world than he takes out.",
+            "autor": "George Bernard Shaw",
+            "genero": "Life",
+            "timestamp": "2025-01-03 14:34:48"
+        },
+        {
+            "texto": "I like the dreams of the future better than the history of the past.",
+            "autor": "Thomas Jefferson",
+            "genero": "Life",
+            "timestamp": "2025-01-03 14:34:39"
+        },
+        {
+            "texto": "Friendship takes time and energy if it's going to work.",
+            "autor": "Betty White",
+            "genero": "Life",
+            "timestamp": "2025-01-03 14:34:37"
+        },
+        {
+            "texto": "You have a brain and mind of your own. Use it, and reach your own decisions.",
+            "autor": "Napoleon Hill",
+            "genero": "Life",
+            "timestamp": "2025-01-03 14:18:47"
+        }
+    ]
+}
