+++ requirements.txt
@@ -1,0 +1,7 @@
+certifi==2024.12.14
+charset-normalizer==3.4.1
+idna==3.10
+requests==2.32.3
+schedule==1.2.2
+tk==0.1.0
+urllib3==2.3.0
