+++ ultimo_estado.json
@@ -1,0 +1,1 @@
+{"categoria": "Life", "citacao": {"texto": "If you don't know where you're going, you will probably end up somewhere else.", "autor": "Laurence J. Peter", "genero": "Life"}, "tema_escuro": false}
