+++ upgrades.md
@@ -1,0 +1,2 @@
+1 - Possibilidade de baixar imagens com as citações em png
+2 - Ao mudar novas citações, o usuário mudar a cor de fundo de acordo com a categoria
