--- upgrades.py
@@ -1,2 +1,0 @@
-# Possibilidade de baixar imagens com as citações em png
-# Ao mudar novas citações, o usuário mudar a cor de fundo de acordo com a categoria
